User Defined Mapping Update + Smoothed Map improvements + New Option Menu + IPython Minor Change
-A new widget was added to edit quickly global variables. It is present
in the Mapping Menu, but will be extended for all necessary options
-The smooth map will only display data which are at a minimal distance
from a real point. interpolation will not be extended to the full map
anymore. This distance is fixed by Mapping.Max_Valid_Dist and
computation use cKDTree
-Do Map button is usable only if a proper mapping was done (Mapping
Average or Mapping Measure). It is also set to inactive if a new SQL
query is done.
-User Defined Pattern will use the number of Turns value to evaluate the
number of Repeats, except if Mapping.Use_Number_of_Turn is set to False
(in which case Mapping.Number_of_Turns is ignored. i.e. = 1)
-Min bug fixes