New IO File to import Neuromatic files. Send to Igor was improved
Neuromatic files can be imported (for spike sorting).
Only one channel at the time is supported for now.
A filter must be used to select the good set of sweeps. By default, use
RecordA for channel1 RecordB for channel 2 etc...
Sending a trace  to Igor was improved (multiple signals can be sent)